Fix: ImageEntity의 스태틱 팩터리 메서드 수정
- 정상 동작하지 않고 있어 수정
- static 필드엔 @Value가 적용되지 않아 별도의 클래스로 분리하였음
